requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
onnxruntime>=1.20.0
python-multipart>=0.0.12
aiohttp>=3.11.0
Pillow>=11.0.0
numpy>=1.26.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
opencv-python-headless>=4.10.0